Fixing oblique forests.
