+++ Makefile
@@ -5,7 +5,7 @@
 PYTHON ?= python
 PYTESTS ?= pytest
 CODESPELL_SKIPS ?= "docs/auto_*,*.fif,*.eve,*.gz,*.tgz,*.zip,*.mat,*.stc,*.label,*.w,*.bz2,*.annot,*.sulc,*.log,*.local-copy,*.orig_avg,*.inflated_avg,*.gii,*.pyc,*.doctree,*.pickle,*.inv,*.png,*.edf,*.touch,*.thickness,*.nofix,*.volume,*.defect_borders,*.mgh,lh.*,rh.*,COR-*,FreeSurferColorLUT.txt,*.examples,.xdebug_mris_calc,bad.segments,BadChannels,*.hist,empty_file,*.orig,*.js,*.map,*.ipynb,searchindex.dat,install_mne_c.rst,plot_*.rst,*.rst.txt,c_EULA.rst*,*.html,gdf_encodes.txt,*.svg"
-CODESPELL_DIRS ?= morf/ docs/ examples/ tests/
+CODESPELL_DIRS ?= oblique_forests/ docs/ examples/ tests/
 
 all: clean inplace test
 
@@ -18,7 +18,7 @@ clean-build:
 	rm -rf _build
 	rm -rf _build
 	rm -rf dist
-	rm -rf morf.egg-info
+	rm -rf oblique_forests.egg-info
 
 clean-ctags:
 	rm -f tags
@@ -34,7 +34,7 @@ inplace:
 
 test-coverage:
 	rm -rf coverage .coverage
-	$(PYTESTS) --cov=morf --cov-report html:coverage
+	$(PYTESTS) --cov=oblique_forests --cov-report html:coverage
 
 
 check-manifest:
@@ -49,7 +49,7 @@ reqs:
 flake:
 	@if command -v flake8 > /dev/null; then \
 		echo "Running flake8"; \
-		flake8 --count morf examples tests; \
+		flake8 --count oblique_forests examples tests; \
 	else \
 		echo "flake8 not found, please install it!"; \
 		exit 1; \
@@ -68,7 +68,7 @@ codespell-error:  # running on travis
 	@codespell -i 0 -q 7 -S $(CODESPELL_SKIPS) --ignore-words=ignore_words.txt $(CODESPELL_DIRS)
 
 type-check:
-	mypy ./morf
+	mypy ./oblique_forests
 
 pep:
 	@$(MAKE) -k flake pydocstyle check-manifest codespell-error type-check
