+++ oblique_forests/__init__.py
@@ -1,4 +1,38 @@
-from .oblique_tree import ObliqueTreeClassifier
-from .sporf import ObliqueForestClassifier
-from .morf import Conv2DObliqueForestClassifier
+"""Manifold random forests."""
+import os
+import sys
 
+__version__ = '0.1dev'
+
+# On OSX, we can get a runtime error due to multiple OpenMP libraries loaded
+# simultaneously. This can happen for instance when calling BLAS inside a
+# prange. Setting the following environment variable allows multiple OpenMP
+# libraries to be loaded. It should not degrade performances since we manually
+# take care of potential over-subcription performance issues, in sections of
+# the code where nested OpenMP loops can happen, by dynamically reconfiguring
+# the inner OpenMP runtime to temporarily disable it while under the scope of
+# the outer OpenMP parallel section.
+os.environ.setdefault("KMP_DUPLICATE_LIB_OK", "True")
+
+# Workaround issue discovered in intel-openmp 2019.5:
+# https://github.com/ContinuumIO/anaconda-issues/issues/11294
+os.environ.setdefault("KMP_INIT_AT_FORK", "FALSE")
+
+try:
+    # This variable is injected in the __builtins__ by the build
+    # process. It is used to enable importing subpackages of sklearn when
+    # the binaries are not built
+    __SKLEARN_SETUP__
+except NameError:
+    __SKLEARN_SETUP__ = False
+
+if __SKLEARN_SETUP__:
+    sys.stderr.write('Partial import of sklearn during the build process.\n')
+    # We are not importing the rest of scikit-learn during the build
+    # process, as it may not be compiled yet
+else:
+    from sklearn import __check_build
+
+    __check_build  # avoid flakes unused variable error
+
+    __all__ = ['tree']
