+++ oblique_forests/sporf.py
@@ -1,6 +1,6 @@
 from sklearn.ensemble._forest import ForestClassifier
 
-from .oblique_tree import ObliqueTreeClassifier
+from .tree.oblique_tree import ObliqueTreeClassifier
 
 
 class ObliqueForestClassifier(ForestClassifier):
